Add files via upload
Using the CountryInfo, matplotlib, and pandas libraries, this program allows the user to input the names of five countries (in English) and, based on this information, extracts specific data such as the capital, territorial area, and total population of each. This data is displayed in an orderly manner in the console. Additionally, a pie chart is generated to visually represent the proportion of the world's population corresponding to each entered country.